+++ CHANGELOG.md
@@ -5,8 +5,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
-### Fixed
-- Now melting curves show the proper (inverted) dissociation curve.
 
 ## [1.4.3]
 ### Added
@@ -14,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Moved functions to separate library
+
+### Fixed
+- Now melting curves show the proper (inverted) dissociation curve.
 
 ## [1.4.2]
 ### Changed